server/requirements.txt
Flask==2.2.2
numpy==1.23.2
plac==1.1.3
requests==2.28.1
tensorflow-cpu==2.10.0
tensorflow_hub==0.12.0
tqdm==4.64.1
spacy==2.3.8
gunicorn==20.1.0
sentencepiece==0.1.97
